cupbearer/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cupbearer"
version = "0.1.0"
description = "Cupbearer personal assistant service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn>=0.30.0,<1.0.0",
  "mem0ai>=1.0.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "pytest-cov>=5.0.0,<6.0.0",
  "ruff>=0.6.0,<1.0.0",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
markers = [
  "e2e: end-to-end tests that exercise the running service",
  "twilio_live: tests that call live Twilio APIs and require real credentials"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
